+++ tests/test.nim
@@ -25,3 +25,5 @@ import ./t_lockfree_stack_typestates
 import ./t_backoff
 import ./t_bind_client
 import ./t_manager_cc_surface
+import ./t_unregister_thread
+import ./t_unregister_thread_stress